backend dependant convolution layer model
in the parser, the implementation for convolution is chosen based on a
`backend` flag, which imports either a hls or chisel implmentation of
the layer. still need to make the modules backend independant also.
