+++ fpgaconvnet/models/layers/__init__.py
@@ -8,5 +8,5 @@ from .BatchNormLayer    import BatchNormLayer
 from .InnerProductLayer import InnerProductLayer
 from .PoolingLayer      import PoolingLayer
 from .ReLULayer         import ReLULayer
-from .ConvolutionLayer  import ConvolutionLayer
+from .ConvolutionLayer  import ConvolutionLayerBase
 from .SqueezeLayer      import SqueezeLayer
