+++ fpgaconvnet/models/layers/chisel/__init__.py
@@ -1,0 +1,1 @@
+from .ConvolutionLayer  import ConvolutionLayer
