+++ fpgaconvnet/models/modules/__init__.py
@@ -15,3 +15,4 @@ from .ReLU import ReLU
 from .SlidingWindow import SlidingWindow
 from .Squeeze import Squeeze
 from .Bias import Bias
+from .VectorDot import VectorDot
